feat: Nop unit test and general testing structure
I held off on creating unit tests until this commit because I wasn't
sure about the general structure of them. Still not sure this is the
best way to go. I'm reading from memory, and I've created the memorybus
trait. Surely I should mock the memory. Ah well, this is functional and
its okay for now. Should do a bit of a tech spike though, because the
point of this project is to not compromise the code even slightly. Clean
as possible.